server/requirements.txt
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
flask-jwt-extended
python-dotenv
Flask-CORS
requests